Moved the custom dialog to its own file
